Move the Bugzilla group lookup into BugzillaAuthorize
