+++ services/hackbot-api/app/bugzilla_authorization.py
@@ -3,14 +3,13 @@
 from __future__ import annotations
 
 import asyncio
-from typing import TYPE_CHECKING
 
+import httpx
 from cachetools import TTLCache
 
-if TYPE_CHECKING:
-    from app.bugzilla_client import BugzillaUserClient
-
 AUTHORIZED_GROUP_ID = 9  # bmo-editbugs-team
+
+_REQUEST_TIMEOUT_SECONDS = 30
 
 
 class BugzillaAuthorizer:
@@ -18,13 +17,13 @@ class BugzillaAuthorizer:
 
     def __init__(
         self,
-        client: BugzillaUserClient,
+        url: str,
         authorized_group_id: int,
         *,
         cache_ttl_seconds: int = 300,
         cache_maxsize: int = 4096,
     ) -> None:
-        self._client = client
+        self._rest_url = url.rstrip("/") + "/rest"
         self._authorized_group_id = authorized_group_id
         self._cache: TTLCache[str, bool] = TTLCache(
             maxsize=cache_maxsize,
@@ -45,8 +44,28 @@ class BugzillaAuthorizer:
             if cached is not None:
                 return cached
 
-            authorized = await self._client.is_user_in_group(
-                login, self._authorized_group_id
-            )
+            authorized = await self._is_user_in_group(login, self._authorized_group_id)
             self._cache[login] = authorized
             return authorized
+
+    # TODO: Move this REST call to a shared Bugzilla client library (#6459).
+    async def _is_user_in_group(self, login: str, group_id: int) -> bool:
+        """Return whether a Bugzilla account exists and belongs to a group.
+
+        The ``group_ids`` parameter filters server-side and needs no API key,
+        so the service holds no Bugzilla credential.
+        """
+        async with httpx.AsyncClient(timeout=_REQUEST_TIMEOUT_SECONDS) as client:
+            response = await client.get(
+                f"{self._rest_url}/user",
+                params={
+                    "names": login,
+                    "group_ids": str(group_id),
+                    "include_fields": "name",
+                    # Report an unknown login in ``faults`` instead of failing
+                    # the request, so it maps to "not authorized", not a 500.
+                    "permissive": "1",
+                },
+            )
+        response.raise_for_status()
+        return bool(response.json().get("users"))
