+++ services/hackbot-api/app/routers/webhooks.py
@@ -12,7 +12,6 @@ from app.auth import (
     require_phabricator_signature,
 )
 from app.bugzilla_authorization import AUTHORIZED_GROUP_ID, BugzillaAuthorizer
-from app.bugzilla_client import BugzillaUserClient
 from app.bugzilla_webhook import detect_needinfo_request
 from app.config import settings
 from app.phabricator_authorization import (
@@ -58,8 +57,9 @@ def get_bugzilla_authorizer(request: Request) -> BugzillaAuthorizer:
     """Dependency: lazily create the app-scoped authorizer and its user cache."""
     authorizer = getattr(request.app.state, "bugzilla_authorizer", None)
     if authorizer is None:
-        client = BugzillaUserClient(settings.bugzilla_webhook.url)
-        authorizer = BugzillaAuthorizer(client, AUTHORIZED_GROUP_ID)
+        authorizer = BugzillaAuthorizer(
+            settings.bugzilla_webhook.url, AUTHORIZED_GROUP_ID
+        )
         request.app.state.bugzilla_authorizer = authorizer
     return authorizer
